+++ WhatsNew.md
@@ -1,4 +1,7 @@
 #csstylus - What's new#
+## v 0.1.0 Error management: 
+Now filename error and file syntax eror or not existing package generate a 404 error with a console message. Parameters compressCss and compressStyl are renamed minifyCss an minifyStyl.
+
 ## v 0.0.5 Relative url: Now you can request a file a ```[mountedPath]/[path]/[filename].css``` and even ```[mountedPath]/[path]/[filename]``` from configured directory options.file. The same options are now supported in static and config methods: ```require('csstylus').static(...).config(...)``` 
 
 ## v 0.0.2 Compression options: .config({compressCss:true,compressStyl:false})